# docs/mobile_v1_openapi.yaml
openapi: 3.0.3
info:
  title: FichaYa - Mobile API v1
  version: 1.25.0
  description: |
    API REST para la aplicacion Flutter de control de asistencia.
    Todos los endpoints (excepto auth) requieren header `Authorization: Bearer <token>`.
    Los modulos complementarios Feedback (`/api/v1/feedback`) y SKAP (`/api/skap`)
    usan el mismo Bearer JWT mobile, aunque no cuelgan del prefijo `/api/v1/mobile`.
    La campana de Alertas del home agrupa adelantos, pedidos de mercaderia y feedback
    para mostrar novedades o aprobaciones pendientes sin cambiar los flujos existentes.
  contact:
    name: Equipo Backend
servers:
  - url: http://localhost:5000/api/v1/mobile
    description: Local
  - url: https://control-asistencia-backend-8gle.onrender.com/api/v1/mobile
    description: Produccion

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    # --- Feedback de calle ---
    FeedbackEmpleadoRef:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        nombre:
          type: string
          nullable: true
        legajo:
          type: string
          nullable: true
        dni:
          type: string
          nullable: true
    FeedbackClienteRef:
      type: object
      properties:
        id:
          type: integer
        codigo:
          type: string
          nullable: true
        razon_social:
          type: string
          nullable: true
        nombre_fantasia:
          type: string
          nullable: true
        tipo:
          type: string
          nullable: true
    FeedbackCliente:
      allOf:
        - $ref: '#/components/schemas/FeedbackClienteRef'
        - type: object
          properties:
            sucursal_origen:
              type: integer
              nullable: true
            telefonos:
              type: string
              nullable: true
            movil:
              type: string
              nullable: true
            email:
              type: string
              nullable: true
            domicilio:
              type: string
              nullable: true
            localidad:
              type: string
              nullable: true
            provincia:
              type: string
              nullable: true
    FeedbackMotivo:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
          nullable: true
        sla_dias:
          type: integer
          minimum: 1
          description: Compatibilidad legacy. Para UI nueva usar tiempo_resolucion_valor/unidad.
        sector_responsable_id:
          type: integer
          nullable: true
        sector_responsable_nombre:
          type: string
          nullable: true
        tiempo_resolucion_valor:
          type: integer
          minimum: 1
        tiempo_resolucion_unidad:
          type: string
          enum: [HORAS, DIAS]
        requiere_foto:
          type: boolean
        requiere_observacion:
          type: boolean
    FeedbackMotivoRef:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
    FeedbackItem:
      type: object
      properties:
        id:
          type: integer
        numero:
          type: string
          nullable: true
          example: FB-00000123
        empresa_id:
          type: integer
        estado:
          type: string
          enum: [pendiente, resuelto]
        estado_actual:
          type: string
          enum: [pendiente, resuelto]
        condicion_temporal:
          type: string
          enum: [pendiente_en_termino, pendiente_vencido, resuelto_en_termino, resuelto_fuera_termino]
        descripcion:
          type: string
        fecha_vencimiento:
          type: string
          format: date
          description: Compatibilidad legacy. Para UI nueva usar fecha_limite.
        fecha_limite:
          type: string
          nullable: true
          description: Fecha/hora limite calculada por backend en tiempo corrido.
        created_at:
          type: string
          nullable: true
        updated_at:
          type: string
          nullable: true
        resuelto_at:
          type: string
          nullable: true
        resuelto_en_sla:
          type: boolean
          nullable: true
        resolucion_descripcion:
          type: string
          nullable: true
        evidencia:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/FeedbackEvidence'
        dias_restantes:
          type: integer
          nullable: true
        minutos_restantes:
          type: integer
          nullable: true
        sector_origen:
          type: object
          properties:
            id:
              type: integer
              nullable: true
            nombre:
              type: string
              nullable: true
        sucursal:
          type: object
          properties:
            id:
              type: integer
              nullable: true
            nombre:
              type: string
              nullable: true
        sector_responsable:
          type: object
          properties:
            id:
              type: integer
              nullable: true
            nombre:
              type: string
              nullable: true
        responsable:
          $ref: '#/components/schemas/FeedbackEmpleadoRef'
        empleado:
          $ref: '#/components/schemas/FeedbackEmpleadoRef'
        jefe_directo:
          $ref: '#/components/schemas/FeedbackEmpleadoRef'
        cliente:
          $ref: '#/components/schemas/FeedbackClienteRef'
        motivo:
          $ref: '#/components/schemas/FeedbackMotivoRef'
        resuelto_por:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/FeedbackEmpleadoRef'
    FeedbackEvidence:
      type: object
      nullable: true
      properties:
        filename:
          type: string
          example: foto_cliente.png
        mime_type:
          type: string
          enum: [image/jpeg, image/png, image/webp]
          example: image/png
        size_bytes:
          type: integer
          maximum: 8388608
          example: 245120
        url:
          type: string
          example: /media/feedback/evidencias/123
    FeedbackCreateRequest:
      type: object
      required: [cliente_id, motivo_id]
      properties:
        cliente_id:
          type: integer
        motivo_id:
          type: integer
        descripcion:
          type: string
          description: Obligatoria solo cuando el motivo tiene requiere_observacion=true.
    FeedbackCreateMultipartRequest:
      type: object
      required: [cliente_id, motivo_id]
      properties:
        cliente_id:
          type: integer
          description: Tambien se acepta como string en form-data.
        motivo_id:
          type: integer
          description: Tambien se acepta como string en form-data.
        descripcion:
          type: string
          description: Obligatoria solo cuando el motivo tiene requiere_observacion=true.
        foto:
          type: string
          format: binary
          description: Evidencia opcional u obligatoria segun requiere_foto del motivo. JPG, PNG o WebP, maximo 8 MB. El backend tambien acepta los nombres evidencia_file, evidencia o foto_file.
    FeedbackResolveRequest:
      type: object
      required: [resolucion_descripcion]
      properties:
        resolucion_descripcion:
          type: string
    FeedbackMotivosResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackMotivo'
        total:
          type: integer
    FeedbackClientesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackCliente'
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
    FeedbackListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackItem'
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
    FeedbackMutationResponse:
      type: object
      properties:
        ok:
          type: boolean
        feedback:
          $ref: '#/components/schemas/FeedbackItem'
    FeedbackDetailResponse:
      type: object
      properties:
        feedback:
          $ref: '#/components/schemas/FeedbackItem'
    FeedbackDashboardResponse:
      type: object
      properties:
        resumen:
          type: object
          properties:
            total:
              type: integer
            resueltos:
              type: integer
            pendientes:
              type: integer
            en_proceso:
              type: integer
              deprecated: true
              description: Compatibilidad legacy; en Feedback por Sector siempre debe interpretarse como 0.
            vencidos:
              type: integer
            resueltos_en_sla:
              type: integer
            resueltos_fuera_sla:
              type: integer
            motivos_distintos:
              type: integer
            clientes_distintos:
              type: integer
            empleados_con_carga:
              type: integer
        top_motivos:
          type: array
          items:
            type: object
            properties:
              motivo_id:
                type: integer
              motivo_nombre:
                type: string
              total:
                type: integer
              resueltos:
                type: integer
        ranking:
          type: array
          items:
            type: object
            properties:
              empleado_id:
                type: integer
              legajo:
                type: string
                nullable: true
              apellido:
                type: string
                nullable: true
              nombre:
                type: string
                nullable: true
              total:
                type: integer
        personal:
          type: object
          nullable: true
          properties:
            empleado_id:
              type: integer
            total_cargados:
              type: integer
            posicion_ranking:
              type: integer
              nullable: true
            total_personal_activo:
              type: integer
            promedio_por_empleado:
              type: number
            porcentaje_sobre_total:
              type: number
        totales:
          type: object
          properties:
            empleados_activos:
              type: integer
            empleados_con_carga:
              type: integer
        empleado:
          type: object
          properties:
            id:
              type: integer
            nombre:
              type: string
              nullable: true
            apellido:
              type: string
              nullable: true
            legajo:
              type: string
              nullable: true
            empresa_id:
              type: integer

    # --- SKAP ---
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/feedback/motivos:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    get:
      tags: [Feedback]
      summary: Motivos activos de feedback
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Motivos activos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackMotivosResponse'
        '401':
          description: Token invalido o vencido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/feedback/clientes:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    get:
      tags: [Feedback]
      summary: Clientes activos importados por CSV
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          required: false
          description: Busca clientes activos por id, sucursal, codigo, razon social, nombre de fantasia o negocio, telefonos, movil, email, domicilio o direccion, localidad, provincia o tipo. La coincidencia es parcial, no distingue mayusculas y admite varias palabras.
          example: Siempre Viva 123
        - name: search
          in: query
          schema:
            type: string
          required: false
          description: Alias de q.
        - name: query
          in: query
          schema:
            type: string
          required: false
          description: Alias de q.
        - name: cliente_q
          in: query
          schema:
            type: string
          required: false
          description: Alias de q, tambien usado por el panel web.
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: Clientes paginados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackClientesResponse'
  /api/v1/feedback/historial:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    get:
      tags: [Feedback]
      summary: Historial del sector origen del empleado autenticado
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, resuelto, vencido]
          required: false
        - name: q
          in: query
          schema:
            type: string
          description: "Misma logica que historial: cliente, motivo, descripcion, resolucion, estado o nombres/legajos de participantes. Admite varias palabras."
          required: false
      responses:
        '200':
          description: Feedbacks originados en el sector del empleado autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackListResponse'
        '401':
          description: Token invalido o empleado inactivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/feedback/bandeja:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    get:
      tags: [Feedback]
      summary: Bandeja del responsable autenticado
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, resuelto, vencido]
          required: false
        - name: q
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Feedbacks asignados a sectores donde el empleado autenticado es responsable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackListResponse'
  /api/v1/feedback/dashboard:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    get:
      tags: [Feedback]
      summary: Dashboard de feedback para el empleado autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Totales, motivos principales, ranking y posicion personal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackDashboardResponse'
  /api/v1/feedback:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    post:
      tags: [Feedback]
      summary: Crea feedback de calle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FeedbackCreateMultipartRequest'
            encoding:
              foto:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '201':
          description: Feedback creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackMutationResponse'
        '400':
          description: Payload invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Empleado sin permiso o configuracion incompleta de sector/responsable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/feedback/{feedback_id}:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    get:
      tags: [Feedback]
      summary: Detalle de feedback
      security:
        - BearerAuth: []
      parameters:
        - name: feedback_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Feedback visible para empleado del sector origen o responsable asignado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackDetailResponse'
        '403':
          description: Sin permiso para ver el feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Feedback no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/feedback/{feedback_id}/tomar:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    post:
      tags: [Feedback]
      summary: Endpoint legacy para tomar feedback
      description: Compatible con versiones anteriores. Valida permisos pero no cambia el estado; el contrato vigente solo usa pendiente/resuelto.
      security:
        - BearerAuth: []
      parameters:
        - name: feedback_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Feedback actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackMutationResponse'
        '403':
          description: Solo el responsable asignado puede tomarlo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Feedback no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/feedback/{feedback_id}/resolver:
    servers:
      - url: http://localhost:5000
        description: Local root
      - url: https://control-asistencia-backend-8gle.onrender.com
        description: Produccion root
    post:
      tags: [Feedback]
      summary: Resuelve feedback con descripcion de gestion
      security:
        - BearerAuth: []
      parameters:
        - name: feedback_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackResolveRequest'
      responses:
        '200':
          description: Feedback resuelto; incluye resuelto_at y resuelto_en_sla
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackMutationResponse'
        '400':
          description: Falta descripcion de resolucion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Solo el responsable asignado puede resolverlo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---- SKAP (prefijo complementario) ----
...